code/middle/valid_sudoku/main_test.go
package valid_sudoku

import (
	"github.com/stretchr/testify/require"
	"testing"
)

func TestName(t *testing.T) {
	t.Run("test1", func(t *testing.T) {
		board := [][]byte{
			[]byte("53..7...."),
			[]byte("6..195..."),
			[]byte(".98....6."),
			[]byte("8...6...3"),
			[]byte("4..8.3..1"),
			[]byte("7...2...6"),
			[]byte(".6....28."),
			[]byte("...419..5"),
			[]byte("....8..79"),
		}
		require.Equal(t, true, isValidSudoku(board))
	})

	t.Run("test2", func(t *testing.T) {
		board := [][]byte{
			[]byte("83..7...."),
			[]byte("6..195..."),
			[]byte(".98....6."),
			[]byte("8...6...3"),
			[]byte("4..8.3..1"),
			[]byte("7...2...6"),
			[]byte(".6....28."),
			[]byte("...419..5"),
			[]byte("....8..79"),
		}
		require.Equal(t, false, isValidSudoku(board))
	})
}
